backend/requirements/base.txt
# Django
Django>=5.0,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-environ>=0.11.0
django-filter>=23.0.0

# Database
psycopg2-binary>=2.9.0

# Cache & Queue
redis>=5.0.0
celery>=5.3.0
django-redis>=5.4.0

# Channels (WebSocket)
channels>=4.0.0
channels-redis>=4.1.0

# HTTP Client
httpx>=0.25.0
openai>=1.0.0

# Document Processing
markdown>=3.5.0
python-frontmatter>=1.0.0

# Math & Science
sympy>=1.12.0

# Security
cryptography>=41.0.0

# OAuth
requests-oauthlib>=1.3.0
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
